refactor: rename setInteger -> setLong
Consistent with how setDouble() accepts a Double even though
the Weaviate's data type is number.
PropertyBuilder deals with Java data types so the methods
should be named accordingly.

+++ src/main/java/io/weaviate/client6/v1/internal/orm/PojoBuilder.java
@@ -125,8 +125,7 @@ final class PojoBuilder<PropertiesT extends Record> implements PropertiesBuilder
   }
 
   @Override
-  // TODO: rename to setLong
-  public void setInteger(String propertyName, Long value) {
+  public void setLong(String propertyName, Long value) {
     if (isType(propertyName, short.class, Short.class)) {
       setValue(propertyName, value.shortValue());
     } else if (isType(propertyName, int.class, Integer.class)) {
